// web-application/src/types-openapi/api/models/ListPriceTableRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListPriceTableRequest
 */
export interface ListPriceTableRequest {
    /**
     * 
     * @type {number}
     * @memberof ListPriceTableRequest
     */
    id?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ListPriceTableRequest
     */
    name?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof ListPriceTableRequest
     */
    isActive?: boolean | null;
}

/**
 * Check if a given object implements the ListPriceTableRequest interface.
 */
export function instanceOfListPriceTableRequest(value: object): value is ListPriceTableRequest {
    return true;
}

export function ListPriceTableRequestFromJSON(json: any): ListPriceTableRequest {
    return ListPriceTableRequestFromJSONTyped(json, false);
}

export function ListPriceTableRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListPriceTableRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'isActive': json['isActive'] == null ? undefined : json['isActive'],
    };
}

export function ListPriceTableRequestToJSON(json: any): ListPriceTableRequest {
    return ListPriceTableRequestToJSONTyped(json, false);
}

export function ListPriceTableRequestToJSONTyped(value?: ListPriceTableRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'isActive': value['isActive'],
    };
}