// web-application/src/types-openapi/api/models/ListPriceTableResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListPriceTableResponse
 */
export interface ListPriceTableResponse {
    /**
     * 
     * @type {number}
     * @memberof ListPriceTableResponse
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof ListPriceTableResponse
     */
    name: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof ListPriceTableResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ListPriceTableResponse
     */
    effectiveFrom?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof ListPriceTableResponse
     */
    effectiveTo?: Date | null;
}

/**
 * Check if a given object implements the ListPriceTableResponse interface.
 */
export function instanceOfListPriceTableResponse(value: object): value is ListPriceTableResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function ListPriceTableResponseFromJSON(json: any): ListPriceTableResponse {
    return ListPriceTableResponseFromJSONTyped(json, false);
}

export function ListPriceTableResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListPriceTableResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'isActive': json['isActive'] == null ? undefined : json['isActive'],
        'effectiveFrom': json['effectiveFrom'] == null ? undefined : (new Date(json['effectiveFrom'])),
        'effectiveTo': json['effectiveTo'] == null ? undefined : (new Date(json['effectiveTo'])),
    };
}

export function ListPriceTableResponseToJSON(json: any): ListPriceTableResponse {
    return ListPriceTableResponseToJSONTyped(json, false);
}

export function ListPriceTableResponseToJSONTyped(value?: ListPriceTableResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'isActive': value['isActive'],
        'effectiveFrom': value['effectiveFrom'] === null ? null : ((value['effectiveFrom'] as any)?.toISOString()),
        'effectiveTo': value['effectiveTo'] === null ? null : ((value['effectiveTo'] as any)?.toISOString()),
    };
}